Refactor extensions code to prepare for devtools extensions
