+++ packages/devtools_app/lib/src/extensions/runtime/README.md
@@ -1,0 +1,6 @@
+## What are runtime extensions?
+
+Runtime extensions are tooling extensions that are loaded into DevTools at runtime.
+These extensions are provided by pub packages, and can be loaded into DevTools dynamically.
+
+This feature is under construction - more documenation to come.
