+++ packages/devtools_app/lib/src/extensions/static/README.md
@@ -1,0 +1,5 @@
+## What are static extensions?
+
+Static extensions are extension points that are predefined at build time.
+These extension points allow for diverging features or code depending on
+where DevTools is built from (i.e. building for the Dart SDK or for google3).
