+++ packages/devtools_app/test/extentsion_points/extensions_external_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_app/src/extension_points/extensions_external.dart';
+import 'package:devtools_app/src/extensions/static/extension_points/extension_points_external.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
