+++ packages/devtools_app/test/shared/about_dialog_test.dart
@@ -3,8 +3,8 @@
 // found in the LICENSE file.
 
 import 'package:devtools_app/devtools.dart' as devtools;
-import 'package:devtools_app/src/extension_points/extensions_base.dart';
-import 'package:devtools_app/src/extension_points/extensions_external.dart';
+import 'package:devtools_app/src/extensions/static/extension_points_base.dart';
+import 'package:devtools_app/src/extensions/static/extension_points_external.dart';
 import 'package:devtools_app/src/framework/about_dialog.dart';
 import 'package:devtools_app/src/framework/release_notes/release_notes.dart';
 import 'package:devtools_app/src/service/service_manager.dart';
